Added files via upload
Basic layout in my mind, utilities should house all rough stuff like reading and writing of the account formats and everything that is reused several times as a function (keep everything modular and simple). It is somewhat similar to what was shown in the uml picture sent to me, basically its using controller as a more of a powerhouse for everything, and splitting account into a account "processor" and account as a data structure style.